feat(kookie): Update DTO from entities

+++ api/src/main/java/com/api/kookie/core/profile/ProfileService.java
@@ -12,7 +12,7 @@ public interface ProfileService {
 
     ProfileDTO getByProfileId(Integer profileId);
 
-    Boolean addFavorite(Integer profileId, Integer recetteId);
+    Boolean addFavorite(Integer profileId, Long recetteId);
 
     List<RecetteDTO> getFavorites(Integer profileId);
 }
